Split common.js and popup.js into modules
