+++ src/common/common.js
@@ -1,0 +1,4 @@
+"use strict";
+
+// by default translate whole site
+import "/common/modules/Localiser.js";
