+++ src/common/modules/Logger.js
@@ -1,6 +1,11 @@
 import {MESSAGE_LEVEL} from "/common/modules/internal/MessageLevel.js";
 import {ADDON_NAME_SHORT} from "/common/modules/GlobalConsts.js";
 
+import * as AddonSettings from "/common/modules/AddonSettings.js";
+
+// redirect module
+export {MESSAGE_LEVEL} from "/common/modules/internal/MessageLevel.js";
+
 let debugMode = null;
 
 const MESSAGE_LEVEL_NAME = Object.freeze({
