+++ src/popup/modules/QrLibKjua.js
@@ -1,5 +1,7 @@
 /* globals kjua */
 
+import * as Logger from "/common/modules/Logger.js";
+
 /**
  * The type of QR code this library generates.
  *
