+++ src/popup/modules/QrLibQrGen.js
@@ -1,3 +1,5 @@
+import * as Logger from "/common/modules/Logger.js";
+
 /* globals qrcodegen */
 const QRC = qrcodegen.QrCode;
