+++ src/popup/qrcode.html
@@ -10,7 +10,7 @@
 		<script src="lib/qrcodegen.js" charset="utf-8"></script>
 
 		<script defer src="../common/lib/lodash/lodash.js" charset="utf-8"></script>
-		<script defer src="../common/common.js" charset="utf-8"></script>
+		<script defer src="../common/common.js" type="module" charset="utf-8"></script>
 		<script defer src="qrcode.js" type="module" charset="utf-8"></script>
 	</head>
 	<body class="center-container">
